Lyne-Hollick recursive digital filter

Lyne-Hollick recursive digital filter

filterLH(streamflow, alpha = 0.925, passes = 3)

Arguments

streamflow

vector - streamflow values

alpha

numeric - filter parameter (defaults to 0.925 following Nathan and McMahon (1990))

passes

numeric - number of times the filter passes over the data (typically 1-3)

Value

a dataframe with 2 columns and nrow = length of streamflow dataset. Column one contains baseflow in the same units as input, column two contains quickflow in the same units as input.

Examples

streamflow <- GreenRiver[,2]
filterLH(streamflow)
#>      baseflow    quickflow
#> 1    1558.333     0.000000
#> 2    1581.976   478.024479
#> 3    1605.203   864.797370
#> 4    1629.083   640.917002
#> 5    1653.177   456.822953
#> 6    1676.517   383.483167
#> 7    1698.493   271.506658
#> 8    1718.913   291.087409
#> 9    1738.288   501.711634
#> 10   1757.155   332.845484
#> 11   1775.380   314.619941
#> 12   1792.674   247.326287
#> 13   1808.932   231.067880
#> 14   1824.171   235.828900
#> 15   1838.607   291.393197
#> 16   1852.457   247.543187
#> 17   1865.714   214.286279
#> 18   1878.276   191.723915
#> 19   1890.131   189.868669
#> 20   1901.484   288.516389
#> 21   1912.707   307.292689
#> 22   1923.932   486.068155
#> 23   1934.640   525.360281
#> 24   1944.330   375.670200
#> 25   1952.939   257.061210
#> 26   1960.448   209.551832
#> 27   1966.868   263.131739
#> 28   1972.224   167.775836
#> 29   1976.541   383.458822
#> 30   1979.786   660.213548
#> 31   1981.844   478.155834
#> 32   1982.581   317.419079
#> 33   1964.091   305.909351
#> 34   1944.102   255.898399
#> 35   1924.569   245.430701
#> 36   1907.102   202.898055
#> 37   1891.462   198.538438
#> 38   1877.796   152.203717
#> 39   1865.440   174.559762
#> 40   1852.082   177.917648
#> 41   1840.089   129.910970
#> 42   1829.541   150.459495
#> 43   1818.137   151.863305
#> 44   1805.823   164.176546
#> 45   1792.481   197.518564
#> 46   1777.984   202.016099
#> 47   1762.747   197.253300
#> 48   1747.159   202.841405
#> 49   1731.523   198.477195
#> 50   1716.241   193.759130
#> 51   1701.341   188.658519
#> 52   1686.856   183.144344
#> 53   1672.006   197.993886
#> 54   1656.763   193.236633
#> 55   1642.312   177.688252
#> 56   1627.905   192.095408
#> 57   1612.284   237.716319
#> 58   1595.396   224.603791
#> 59   1579.618   180.382477
#> 60   1564.962   215.038488
#> 61   1548.998   251.002167
#> 62   1532.176   217.824386
#> 63   1516.947   173.053390
#> 64   1502.885   207.115152
#> 65   1487.548   252.451600
#> 66   1470.753   249.247213
#> 67   1452.396   277.604127
#> 68   1432.305   337.694782
#> 69   1410.191   379.808581
#> 70   1385.829   384.171217
#> 71   1359.084   420.915636
#> 72   1329.674   530.326097
#> 73   1297.343   442.657459
#> 74   1269.510   280.489790
#> 75   1246.585   423.414638
#> 76   1221.359   388.640656
#> 77   1193.865   366.135000
#> 78   1165.800   354.200000
#> 79   1146.000   154.000000
#> 80   1140.000     0.000000
#> 81   1140.028    19.971875
#> 82   1140.179    49.821406
#> 83   1140.630   109.370229
#> 84   1141.659   188.340857
#> 85   1143.515   206.485059
#> 86   1146.306   253.693525
#> 87   1150.249   389.751431
#> 88   1155.870   564.129875
#> 89   1163.783   686.217013
#> 90   1174.535   895.465188
#> 91   1186.300   843.699608
#> 92   1196.866   683.134155
#> 93   1206.035   633.965158
#> 94   1213.668   426.331718
#> 95   1219.810   190.190166
#> 96   1224.700   175.299888
#> 97   1228.502   291.497666
#> 98   1231.221   408.778870
#> 99   1232.803   377.197469
#> 100  1233.331   106.669430
#> 101  1230.000     0.000000
#> 102  1230.084    59.915625
#> 103  1230.662   239.337656
#> 104  1232.827   647.173420
#> 105  1238.066   871.934420
#> 106  1246.947   733.053419
#> 107  1258.697   611.303062
#> 108  1272.218   507.782459
#> 109  1286.642   403.357616
#> 110  1301.352   368.648036
#> 111  1316.208   473.791631
#> 112  1331.472   498.527542
#> 113  1346.420   503.579916
#> 114  1360.120   509.880144
#> 115  1372.501   497.498645
#> 116  1383.512   396.487532
#> 117  1393.114   426.885721
#> 118  1401.279   368.721360
#> 119  1407.980   402.020283
#> 120  1413.191   336.808713
#> 121  1416.889   343.111206
#> 122  1419.049   340.951035
#> 123  1417.174   322.825897
#> 124  1396.998   253.001524
#> 125  1380.133   179.866512
#> 126  1371.631    38.369202
#> 127  1368.353   151.647449
#> 128  1364.213   225.787064
#> 129  1358.928   151.071669
#> 130  1353.000    77.000000
#> 131  1350.000     0.000000
#> 132  1350.141    99.859375
#> 133  1350.837   209.163281
#> 134  1352.665   357.334893
#> 135  1356.113   413.887031
#> 136  1361.349   438.651028
#> 137  1368.339   491.661373
#> 138  1375.958   424.041762
#> 139  1382.948   477.052264
#> 140  1389.093   510.906743
#> 141  1394.270   535.729516
#> 142  1398.351   441.648921
#> 143  1401.222   408.778290
#> 144  1402.794   357.206285
#> 145  1396.581   393.418532
#> 146  1376.411   423.588510
#> 147  1353.650   466.349567
#> 148  1328.160   441.840115
#> 149  1301.944   328.056094
#> 150  1281.831   188.168750
#> 151  1272.250    57.750000
#> 152  1270.000     0.000000
#> 153  1270.323   229.676563
#> 154  1272.019   547.981328
#> 155  1276.256   653.743761
#> 156  1283.348   606.651621
#> 157  1292.942   637.058494
#> 158  1304.789   705.210670
#> 159  1318.907   781.092503
#> 160  1335.620  1024.379939
#> 161  1356.005  1493.994677
#> 162  1381.120  1438.880066
#> 163  1410.280  1079.719553
#> 164  1441.543   888.456603
#> 165  1473.431   796.569049
#> 166  1505.027   584.973309
#> 167  1535.422   424.578417
#> 168  1563.999   396.000721
#> 169  1590.679   369.321426
#> 170  1615.626   374.374334
#> 171  1639.161   440.838747
#> 172  1662.046   697.954206
#> 173  1685.580   974.419502
#> 174  1710.819   979.181199
#> 175  1737.963  1012.036595
#> 176  1766.741   943.258912
#> 177  1796.628   843.372363
#> 178  1826.929   733.070777
#> 179  1856.923   583.076710
#> 180  1885.937   484.062917
#> 181  1913.808   616.191700
#> 182  1941.009   668.990810
#> 183  1967.948   652.052414
#> 184  1994.932   825.068266
#> 185  2022.830  1117.170196
#> 186  2052.970  1397.029889
#> 187  2086.280  1433.720296
#> 188  2122.550  1257.450412
#> 189  2160.371   959.628773
#> 190  2197.003   752.996566
#> 191  2230.700   779.299777
#> 192  2261.511   898.488690
#> 193  2289.457   830.543441
#> 194  2314.529   795.470695
#> 195  2336.726   653.274499
#> 196  2356.124   423.875653
#> 197  2373.084   286.915826
#> 198  2388.112   351.887751
#> 199  2401.453   538.546912
#> 200  2413.111   566.889455
#> 201  2423.046   416.953598
#> 202  2431.233   288.767298
#> 203  2437.845   172.155253
#> 204  2443.397    66.602984
#> 205  2448.448   101.551510
#> 206  2453.426   196.573741
#> 207  2458.721   241.279222
#> 208  2464.721   365.279404
#> 209  2472.725  1067.275488
#> 210  2485.576  1764.423712
#> 211  2506.441  2413.558778
#> 212  2536.843  2323.156514
#> 213  2575.601  1994.399258
#> 214  2619.995  1670.005023
#> 215  2667.486  1422.513928
#> 216  2716.741  1613.258968
#> 217  2768.353  2081.646862
#> 218  2823.745  2206.255115
#> 219  2883.114  2016.886303
#> 220  2945.833  2234.167316
#> 221  3011.965  2378.034629
#> 222  3081.884  2488.115592
#> 223  3155.789  2654.211090
#> 224  3234.221  3005.779426
#> 225  3320.247  4779.752887
#> 226  3419.319  5530.680631
#> 227  3534.073  5465.926854
#> 228  3662.822  5257.177628
#> 229  3804.327  6295.673072
#> 230  3959.437  6340.563200
#> 231  4126.637  5533.362898
#> 232  4302.482  5897.517598
#> 233  4486.760  6713.239552
#> 234  4681.914  7318.085926
#> 235  4890.057  7909.943372
#> 236  5106.436  8393.563934
#> 237  5314.897  7885.103472
#> 238  5504.409  5795.590876
#> 239  5674.468  3865.532349
#> 240  5826.202  2443.797874
#> 241  5963.123  1676.877096
#> 242  6092.770  4307.230064
#> 243  6227.421  6572.578903
#> 244  6365.745  7934.255057
#> 245  6494.829  9605.171405
#> 246  6608.862 10491.138226
#> 247  6705.172 10194.828046
#> 248  6781.060  8918.939750
#> 249  6834.241  6965.759315
#> 250  6863.322  4036.677757
#> 251  6811.700  3048.299564
#> 252  6643.595  1266.404934
#> 253  6544.562  1275.437766
#> 254  6439.765  2290.235085
#> 255  6321.405  3078.594644
#> 256  6185.115  3364.884847
#> 257  6028.986  3261.013876
#> 258  5852.616  3447.383868
#> 259  5654.689  4045.310586
#> 260  5432.528  4427.472301
#> 261  5184.081  4415.919320
#> 262  4909.707  3820.292538
#> 263  4656.659  2813.340965
#> 264  4483.685  1626.314557
#> 265  4369.255  1310.745467
#> 266  4269.059  1260.941045
#> 267  4166.562  1533.437954
#> 268  4054.847  1745.153405
#> 269  3931.409  2698.591072
#> 270  3790.786  4179.214428
#> 271  3627.348  3972.652264
#> 272  3440.801  3099.199134
#> 273  3269.680  1860.319971
#> 274  3143.573  1376.426995
#> 275  3049.809  1030.191346
#> 276  2965.976  1314.024428
#> 277  2877.372  1152.627761
#> 278  2796.078   933.921903
#> 279  2721.166   988.834490
#> 280  2639.895  1790.104854
#> 281  2548.940  1811.059977
#> 282  2453.732  1206.268219
#> 283  2370.115   939.884561
#> 284  2298.773   891.226552
#> 285  2231.377   838.623300
#> 286  2168.650   771.349513
#> 287  2112.595   667.404879
#> 288  2064.968   555.032302
#> 289  2024.019   495.980867
#> 290  1986.237   473.763099
#> 291  1951.472   418.527675
#> 292  1924.024   285.975865
#> 293  1906.513   163.487421
#> 294  1896.500    93.499915
#> 295  1890.135    69.864773
#> 296  1885.687    44.313268
#> 297  1882.079    57.921438
#> 298  1878.057   111.942645
#> 299  1873.431   116.568570
#> 300  1868.021   171.979371
#> 301  1861.655   168.344777
#> 302  1854.205   205.794711
#> 303  1845.534   224.465695
#> 304  1835.468   264.532078
#> 305  1824.390   145.609507
#> 306  1814.206   115.794062
#> 307  1804.088   605.912500
#> 308  1791.501   298.499324
#> 309  1778.218   161.782377
#> 310  1765.911   154.089056
#> 311  1758.282    41.717899
#> 312  1754.884    55.115634
#> 313  1750.756   329.243802
#> 314  1745.325   144.674804
#> 315  1741.125    28.875000
#> 316  1740.000     0.000000
#> 317  1740.007    19.993007
#> 318  1740.017    19.983453
#> 319  1740.000     0.000000
#> 320  1741.069   758.931250
#> 321  1744.382   185.618125
#> 322  1749.084   210.915871
#> 323  1753.799   166.200621
#> 324  1758.105   111.895465
#> 325  1762.022    47.978393
#> 326  1765.957   284.042513
#> 327  1770.557   259.442762
#> 328  1775.584   134.416128
#> 329  1780.352   239.648265
#> 330  1784.661   265.338918
#> 331  1788.464   241.536223
#> 332  1791.709   188.290950
#> 333  1794.363   105.636896
#> 334  1796.453    73.546822
#> 335  1798.058    91.941695
#> 336  1799.220   100.780237
#> 337  1799.942    70.058150
#> 338  1800.295    29.705186
#> 339  1800.406    29.593673
#> 340  1798.290    41.710080
#> 341  1796.920     3.080154
#> 342  1796.594    53.405910
#> 343  1796.035    43.965215
#> 344  1795.208    34.791876
#> 345  1794.124    55.875718
#> 346  1792.716    67.284142
#> 347  1790.959    39.040915
#> 348  1790.000     0.000000
#> 349  1790.070    49.929688
#> 350  1790.693   299.307422
#> 351  1793.176   736.824204
#> 352  1799.009   860.990802
#> 353  1809.026  1110.973712
#> 354  1823.143   856.856737
#> 355  1840.239   739.761206
#> 356  1858.985   591.014873
#> 357  1878.518   561.481834
#> 358  1898.980   951.019916
#> 359  1921.050   778.950201
#> 360  1944.315   605.685206
#> 361  1967.923   642.076803
#> 362  1991.552   548.447994
#> 363  2014.883   475.117449
#> 364  2037.489   382.510841
#> 365  2059.012   310.988313
#> 366  2079.275   310.725354
#> 367  2098.951   731.049404
#> 368  2120.095  1289.905141
#> 369  2143.825   696.175053
#> 370  2168.804   641.195636
#> 371  2193.772   686.228450
#> 372  2217.647   822.352727
#> 373  2239.460  1410.539760
#> 374  2259.013  1030.986717
#> 375  2276.089   653.911125
#> 376  2290.578   469.422288
#> 377  2302.589   347.410771
#> 378  2312.507   257.492721
#> 379  2320.875   159.125304
#> 380  2328.168   131.831959
#> 381  2334.645   115.355497
#> 382  2340.435    89.565320
#> 383  2345.680    64.319796
#> 384  2350.532    79.467686
#> 385  2355.131    94.869094
#> 386  2359.543    70.457035
#> 387  2363.742    66.257521
#> 388  2367.722    72.277551
#> 389  2371.928   358.071878
#> 390  2376.781   103.218620
#> 391  2382.330   427.669821
#> 392  2389.188   540.811799
#> 393  2397.733   392.267463
#> 394  2407.201   312.798975
#> 395  2416.453   223.546981
#> 396  2424.926   195.074317
#> 397  2432.625   227.375246
#> 398  2439.556   220.444162
#> 399  2445.719   184.281028
#> 400  2451.115   188.885039
#> 401  2455.744   214.255593
#> 402  2459.600   220.400036
#> 403  2462.663   257.337448
#> 404  2464.901   255.098805
#> 405  2466.277   233.722644
#> 406  2466.751   253.248835
#> 407  2453.986   306.014283
#> 408  2439.272   310.728374
#> 409  2422.683   307.316828
#> 410  2404.089   395.910923
#> 411  2383.239   386.760559
#> 412  2360.038   379.962116
#> 413  2334.526   375.474255
#> 414  2306.700   393.300087
#> 415  2276.452   433.548016
#> 416  2244.513   415.487361
#> 417  2215.284   334.716066
#> 418  2191.388   278.611963
#> 419  2171.636   228.364285
#> 420  2157.579   132.420848
#> 421  2146.766   193.233687
#> 422  2134.687   315.312812
#> 423  2120.599   499.400596
#> 424  2103.809   466.190816
#> 425  2084.108   485.891994
#> 426  2061.315   598.684744
#> 427  2035.156   534.844497
#> 428  2005.789   444.211074
#> 429  1973.402   566.598191
#> 430  1937.736   582.263609
#> 431  1903.587   396.412942
#> 432  1871.369   478.630545
#> 433  1835.861   814.138807
#> 434  1795.752   924.248252
#> 435  1750.220   909.780489
#> 436  1699.033   980.966797
#> 437  1642.003   977.996822
#> 438  1579.011   990.989328
#> 439  1509.977  1020.023323
#> 440  1456.255   453.745018
#> 441  1427.573   282.427047
#> 442  1407.106   242.894105
#> 443  1389.844   200.155789
#> 444  1376.453   143.546799
#> 445  1367.652    82.347891
#> 446  1362.597    47.403125
#> 447  1360.375     9.625000
#> 448  1360.000     0.000000
#> 449  1360.028    19.971875
#> 450  1360.235    89.765156
#> 451  1361.184   348.816166
#> 452  1364.000   635.999908
#> 453  1369.945   890.054935
#> 454  1379.954  1080.046083
#> 455  1394.579  1255.421063
#> 456  1414.112  1395.887974
#> 457  1438.424  1351.576230
#> 458  1466.941  1383.059378
#> 459  1499.204  1470.795562
#> 460  1534.985  1485.015047
#> 461  1573.756  1356.244010
#> 462  1614.502  1185.497918
#> 463  1656.273  1183.727055
#> 464  1698.742  1231.258083
#> 465  1741.696  1058.304117
#> 466  1784.429   905.570545
#> 467  1826.129   783.870610
#> 468  1866.528   883.472456
#> 469  1905.908   874.091940
#> 470  1944.358   775.642306
#> 471  1980.819   749.181134
#> 472  2014.404   735.596117
#> 473  2045.179   844.820970
#> 474  2073.180   746.820094
#> 475  2098.425   691.574794
#> 476  2120.941   599.058653
#> 477  2140.767   609.232934
#> 478  2157.942   562.057633
#> 479  2172.502   527.497694
#> 480  2184.483   435.517356
#> 481  2193.978   376.022404
#> 482  2201.289   268.711420
#> 483  2207.011   132.988884
#> 484  2211.834    58.166280
#> 485  2216.281   113.719434
#> 486  2220.880   279.120242
#> 487  2226.244   343.755632
#> 488  2232.112   307.888334
#> 489  2237.670   272.330023
#> 490  2242.637   197.363169
#> 491  2246.980   143.020313
#> 492  2250.696    99.303893
#> 493  2253.852    66.148328
#> 494  2256.605    33.395016
#> 495  2259.110    30.890390
#> 496  2261.455    48.545485
#> 497  2263.734    66.266058
#> 498  2266.071    93.928915
#> 499  2268.613   131.386783
#> 500  2271.513   158.486684
#> 501  2274.914   205.085674
#> 502  2278.964   241.036140
#> 503  2283.742   246.257617
#> 504  2289.174   220.825982
#> 505  2295.050   184.950455
#> 506  2301.159   178.840674
#> 507  2307.393   172.607014
#> 508  2313.712   176.288361
#> 509  2320.138   199.862217
#> 510  2326.789   243.211497
#> 511  2333.960   366.040411
#> 512  2342.014   387.986297
#> 513  2351.200   488.800074
#> 514  2361.655   478.345298
#> 515  2373.263   446.737364
#> 516  2385.497   264.503364
#> 517  2397.818   342.182192
#> 518  2409.884   380.115801
#> 519  2421.236   398.764018
#> 520  2431.567   438.433114
#> 521  2440.825   419.174690
#> 522  2448.954   401.046267
#> 523  2455.907   214.092873
#> 524  2461.802   108.198463
#> 525  2466.973    63.027328
#> 526  2471.728    78.272153
#> 527  2476.251   103.749164
#> 528  2480.817   209.183092
#> 529  2485.732   194.268341
#> 530  2490.987   149.012586
#> 531  2496.463   203.536810
#> 532  2502.603   477.397329
#> 533  2510.541   749.458876
#> 534  2521.057   668.943019
#> 535  2534.577  1125.423209
#> 536  2552.827  1827.173253
#> 537  2577.637  1752.363098
#> 538  2608.582  1571.418491
#> 539  2644.123  1545.877283
#> 540  2682.975  1337.024902
#> 541  2724.245  1485.754819
#> 542  2766.838  1873.161562
#> 543  2807.792  2292.208335
#> 544  2844.753  1605.246535
#> 545  2877.418   852.581971
#> 546  2906.149   413.850633
#> 547  2932.037   317.963398
#> 548  2955.968   354.031768
#> 549  2978.394   361.606026
#> 550  2999.518   300.482404
#> 551  3019.380   310.619854
#> 552  3038.221   391.779286
#> 553  3056.896   723.103503
#> 554  3077.168  1212.831580
#> 555  3100.309   969.690800
#> 556  3125.620   764.380210
#> 557  3151.815   728.185135
#> 558  3178.226   621.774370
#> 559  3204.476   625.524491
#> 560  3230.848   909.152112
#> 561  3259.202  1670.798266
#> 562  3295.349  4194.650703
#> 563  3351.068  7548.931722
#> 564  3441.754 11258.245990
#> 565  3581.128 13818.872252
#> 566  3771.923 11928.076942
#> 567  4000.018  8399.982271
#> 568  4232.866  6267.133624
#> 569  4447.674  6152.325919
#> 570  4642.876  7957.123536
#> 571  4817.869  7882.131077
#> 572  4971.842  5128.158305
#> 573  5104.667  3315.333116
#> 574  5219.024  1970.976140
#> 575  5320.483  1239.516992
#> 576  5413.447  1146.553218
#> 577  5501.491  2518.508602
#> 578  5591.540  4278.460378
#> 579  5691.219  5008.780903
#> 580  5805.061  5994.938571
#> 581  5934.924  6065.076322
#> 582  6081.981  7418.018756
#> 583  6251.081  9848.919395
#> 584  6450.745 11949.255458
#> 585  6688.093 13511.906565
#> 586  6965.914 14134.086069
#> 587  7281.581 13618.418547
#> 588  7626.591 11873.409420
#> 589  7987.733  9512.266808
#> 590  8352.070  8447.929534
#> 591  8714.279  9485.720726
#> 592  9078.536 11221.464261
#> 593  9452.058 12047.942211
#> 594  9838.579 12261.421482
#> 595 10237.422 11862.577937
#> 596 10646.956 12553.043554
#> 597 11068.678 13731.321702
#> 598 11506.331 14593.669133
#> 599 11962.043 14837.957014
#> 600 12435.003 14764.996700
#> 601 12922.733 14677.266924
#> 602 13420.323 14479.677011
#> 603 13901.227 13898.772660
#> 604 14342.073 13557.927281
#> 605 14742.171 13257.828810
#> 606 15100.817 12799.183114
#> 607 15417.276 12582.724120
#> 608 15690.816 11509.183776
#> 609 15920.832 10179.167661
#> 610 16106.983  9293.016687
#> 611 16249.202  8150.797569
#> 612 16347.633  7452.366600
#> 613 16403.640  6396.359834
#> 614 16422.096  4777.903753
#> 615 16063.534  4936.466128
#> 616 15699.766  4400.233652
#> 617 15367.315  4132.685029
#> 618 15030.712  5469.287869
#> 619 14658.114  8641.886379
#> 620 14236.004  9863.996360
#> 621 13754.670 10645.329967
#> 622 13209.209 11090.790581
#> 623 12597.217 11002.783030
#> 624 11919.988  9880.011997
#> 625 11227.640  8672.360298
#> 626 10601.502  7398.497620
#> 627 10074.597  6125.402832
#> 628  9594.159  6205.840900
#> 629  9090.679  6909.321243
#> 630  8582.862  6317.137830
#> 631  8090.932  6309.067925
#> 632  7595.602  6404.397756
#> 633  7104.705  6195.294872
#> 634  6646.979  5553.021483
#> 635  6229.166  5170.834036
#> 636  5842.342  4757.658417
#> 637  5492.667  4217.333424
#> 638  5190.450  3539.549647
#> 639  4933.865  3046.134754
#> 640  4707.962  2752.037572
#> 641  4507.122  2402.878456
#> 642  4334.996  2015.003736
#> 643  4187.834  1762.166201
#> 644  4054.685  1655.314812
#> 645  3928.173  1591.826824
#> 646  3808.025  1491.974945
#> 647  3694.351  1425.648589
#> 648  3586.866  1333.133610
#> 649  3484.450  1295.549848
#> 650  3386.297  1223.702539
#> 651  3296.403  1083.597339
#> 652  3216.246   973.753880
#> 653  3144.996   855.004195
#> 654  3084.590   695.409940
#> 655  3035.097   574.902638
#> 656  2993.754   486.246095
#> 657  2961.220   348.779563
#> 658  2937.806   252.194122
#> 659  2920.196   199.804456
#> 660  2907.644   122.356168
#> 661  2899.750    80.249912
#> 662  2895.270    34.729634
#> 663  2892.409    67.591159
#> 664  2889.121   110.878913
#> 665  2885.250    94.749715
#> 666  2880.727   129.272711
#> 667  2875.327   224.673041
#> 668  2868.864   121.135572
#> 669  2861.940    88.060218
#> 670  2855.611    74.389425
#> 671  2849.579    80.421000
#> 672  2842.922   177.078446
#> 673  2835.218   254.782314
#> 674  2825.873   394.126751
#> 675  2814.421   345.579054
#> 676  2800.654   439.345688
#> 677  2784.611   275.388749
#> 678  2766.696   243.304107
#> 679  2748.723   221.276631
#> 680  2730.767   249.233182
#> 681  2711.295   278.705341
#> 682  2691.070   258.930309
#> 683  2671.292   248.708442
#> 684  2653.153   216.846965
#> 685  2637.598   182.402124
#> 686  2623.619   176.380675
#> 687  2612.561   107.438567
#> 688  2606.688    43.311964
#> 689  2602.872   247.127799
#> 690  2597.690   362.309614
#> 691  2590.716   209.284190
#> 692  2582.225   217.774683
#> 693  2575.128    64.871522
#> 694  2569.717   170.283402
#> 695  2563.512   116.487546
#> 696  2558.984    31.015675
#> 697  2556.470    33.530459
#> 698  2553.720    56.279550
#> 699  2550.599    99.401288
#> 700  2546.857   153.142695
#> 701  2542.397    67.602537
#> 702  2540.000     0.000000
#> 703  2540.014     9.985938
#> 704  2540.195    99.805234
#> 705  2540.683    79.316698
#> 706  2541.286   128.713682
#> 707  2541.784   158.215725
#> 708  2542.123   207.877465
#> 709  2541.944   218.055944
#> 710  2537.338   222.661783
#> 711  2531.424   198.576039
#> 712  2524.378   125.621839
#> 713  2517.391    82.609344
#> 714  2510.662   109.337805
#> 715  2502.904   377.096022
#> 716  2493.613   146.387256
#> 717  2484.863    95.136816
#> 718  2477.073   152.926625
#> 719  2468.460   141.539975
#> 720  2458.986   131.014041
#> 721  2449.746   110.253529
#> 722  2442.023    87.976788
#> 723  2434.875   105.125244
#> 724  2427.057   142.943380
#> 725  2418.354   191.645969
#> 726  2408.380   311.620392
#> 727  2396.510   433.489804
#> 728  2382.202   427.797667
#> 729  2365.339   394.661395
#> 730  2346.029   393.970698
#> 731  2324.294   415.706175
#> 732  2300.000   460.000000