Lyne-Hollick recursive digital filter

Lyne-Hollick recursive digital filter

filterLH(dataset, alpha = 0.925, passes = 3)

Arguments

dataset

dataframe - column 1 contains dates, column 2 contains streamflow values

alpha

numeric - filter parameter (defaults to 0.925 following Nathan and McMahon (1990))

passes

numeric - number of times the filter passes over the data (typically 1-3)

Value

a dataframe with 3 columns and nrow = length of streamflow dataset. Column one contains streamflow in the same units as input, column two contains baseflow in the same units as input, column three contains quickflow in the same units as input.

Examples

filterLH(GreenRiver)
#>        dates streamflow  baseflow    quickflow
#> 1    9/30/21       1870  1558.333     0.000000
#> 2    10/1/21       2060  1581.976   478.024479
#> 3    10/2/21       2470  1605.203   864.797370
#> 4    10/3/21       2270  1629.083   640.917002
#> 5    10/4/21       2110  1653.177   456.822953
#> 6    10/5/21       2060  1676.517   383.483167
#> 7    10/6/21       1970  1698.493   271.506658
#> 8    10/7/21       2010  1718.913   291.087409
#> 9    10/8/21       2240  1738.288   501.711634
#> 10   10/9/21       2090  1757.155   332.845484
#> 11  10/10/21       2090  1775.380   314.619941
#> 12  10/11/21       2040  1792.674   247.326287
#> 13  10/12/21       2040  1808.932   231.067880
#> 14  10/13/21       2060  1824.171   235.828900
#> 15  10/14/21       2130  1838.607   291.393197
#> 16  10/15/21       2100  1852.457   247.543187
#> 17  10/16/21       2080  1865.714   214.286279
#> 18  10/17/21       2070  1878.276   191.723915
#> 19  10/18/21       2080  1890.131   189.868669
#> 20  10/19/21       2190  1901.484   288.516389
#> 21  10/20/21       2220  1912.707   307.292689
#> 22  10/21/21       2410  1923.932   486.068155
#> 23  10/22/21       2460  1934.640   525.360281
#> 24  10/23/21       2320  1944.330   375.670200
#> 25  10/24/21       2210  1952.939   257.061210
#> 26  10/25/21       2170  1960.448   209.551832
#> 27  10/26/21       2230  1966.868   263.131739
#> 28  10/27/21       2140  1972.224   167.775836
#> 29  10/28/21       2360  1976.541   383.458822
#> 30  10/29/21       2640  1979.786   660.213548
#> 31  10/30/21       2460  1981.844   478.155834
#> 32  10/31/21       2300  1982.581   317.419079
#> 33   11/1/21       2270  1964.091   305.909351
#> 34   11/2/21       2200  1944.102   255.898399
#> 35   11/3/21       2170  1924.569   245.430701
#> 36   11/4/21       2110  1907.102   202.898055
#> 37   11/5/21       2090  1891.462   198.538438
#> 38   11/6/21       2030  1877.796   152.203717
#> 39   11/7/21       2040  1865.440   174.559762
#> 40   11/8/21       2030  1852.082   177.917648
#> 41   11/9/21       1970  1840.089   129.910970
#> 42  11/10/21       1980  1829.541   150.459495
#> 43  11/11/21       1970  1818.137   151.863305
#> 44  11/12/21       1970  1805.823   164.176546
#> 45  11/13/21       1990  1792.481   197.518564
#> 46  11/14/21       1980  1777.984   202.016099
#> 47  11/15/21       1960  1762.747   197.253300
#> 48  11/16/21       1950  1747.159   202.841405
#> 49  11/17/21       1930  1731.523   198.477195
#> 50  11/18/21       1910  1716.241   193.759130
#> 51  11/19/21       1890  1701.341   188.658519
#> 52  11/20/21       1870  1686.856   183.144344
#> 53  11/21/21       1870  1672.006   197.993886
#> 54  11/22/21       1850  1656.763   193.236633
#> 55  11/23/21       1820  1642.312   177.688252
#> 56  11/24/21       1820  1627.905   192.095408
#> 57  11/25/21       1850  1612.284   237.716319
#> 58  11/26/21       1820  1595.396   224.603791
#> 59  11/27/21       1760  1579.618   180.382477
#> 60  11/28/21       1780  1564.962   215.038488
#> 61  11/29/21       1800  1548.998   251.002167
#> 62  11/30/21       1750  1532.176   217.824386
#> 63   12/1/21       1690  1516.947   173.053390
#> 64   12/2/21       1710  1502.885   207.115152
#> 65   12/3/21       1740  1487.548   252.451600
#> 66   12/4/21       1720  1470.753   249.247213
#> 67   12/5/21       1730  1452.396   277.604127
#> 68   12/6/21       1770  1432.305   337.694782
#> 69   12/7/21       1790  1410.191   379.808581
#> 70   12/8/21       1770  1385.829   384.171217
#> 71   12/9/21       1780  1359.084   420.915636
#> 72  12/10/21       1860  1329.674   530.326097
#> 73  12/11/21       1740  1297.343   442.657459
#> 74  12/12/21       1550  1269.510   280.489790
#> 75  12/13/21       1670  1246.585   423.414638
#> 76  12/14/21       1610  1221.359   388.640656
#> 77  12/15/21       1560  1193.865   366.135000
#> 78  12/16/21       1520  1165.800   354.200000
#> 79  12/17/21       1300  1146.000   154.000000
#> 80  12/18/21       1140  1140.000     0.000000
#> 81  12/19/21       1160  1140.028    19.971875
#> 82  12/20/21       1190  1140.179    49.821406
#> 83  12/21/21       1250  1140.630   109.370229
#> 84  12/22/21       1330  1141.659   188.340857
#> 85  12/23/21       1350  1143.515   206.485059
#> 86  12/24/21       1400  1146.306   253.693525
#> 87  12/25/21       1540  1150.249   389.751431
#> 88  12/26/21       1720  1155.870   564.129875
#> 89  12/27/21       1850  1163.783   686.217013
#> 90  12/28/21       2070  1174.535   895.465188
#> 91  12/29/21       2030  1186.300   843.699608
#> 92  12/30/21       1880  1196.866   683.134155
#> 93  12/31/21       1840  1206.035   633.965158
#> 94    1/1/22       1640  1213.668   426.331718
#> 95    1/2/22       1410  1219.810   190.190166
#> 96    1/3/22       1400  1224.700   175.299888
#> 97    1/4/22       1520  1228.502   291.497666
#> 98    1/5/22       1640  1231.221   408.778870
#> 99    1/6/22       1610  1232.803   377.197469
#> 100   1/7/22       1340  1233.331   106.669430
#> 101   1/8/22       1230  1230.000     0.000000
#> 102   1/9/22       1290  1230.084    59.915625
#> 103  1/10/22       1470  1230.662   239.337656
#> 104  1/11/22       1880  1232.827   647.173420
#> 105  1/12/22       2110  1238.066   871.934420
#> 106  1/13/22       1980  1246.947   733.053419
#> 107  1/14/22       1870  1258.697   611.303062
#> 108  1/15/22       1780  1272.218   507.782459
#> 109  1/16/22       1690  1286.642   403.357616
#> 110  1/17/22       1670  1301.352   368.648036
#> 111  1/18/22       1790  1316.208   473.791631
#> 112  1/19/22       1830  1331.472   498.527542
#> 113  1/20/22       1850  1346.420   503.579916
#> 114  1/21/22       1870  1360.120   509.880144
#> 115  1/22/22       1870  1372.501   497.498645
#> 116  1/23/22       1780  1383.512   396.487532
#> 117  1/24/22       1820  1393.114   426.885721
#> 118  1/25/22       1770  1401.279   368.721360
#> 119  1/26/22       1810  1407.980   402.020283
#> 120  1/27/22       1750  1413.191   336.808713
#> 121  1/28/22       1760  1416.889   343.111206
#> 122  1/29/22       1760  1419.049   340.951035
#> 123  1/30/22       1740  1417.174   322.825897
#> 124  1/31/22       1650  1396.998   253.001524
#> 125   2/1/22       1560  1380.133   179.866512
#> 126   2/2/22       1410  1371.631    38.369202
#> 127   2/3/22       1520  1368.353   151.647449
#> 128   2/4/22       1590  1364.213   225.787064
#> 129   2/5/22       1510  1358.928   151.071669
#> 130   2/6/22       1430  1353.000    77.000000
#> 131   2/7/22       1350  1350.000     0.000000
#> 132   2/8/22       1450  1350.141    99.859375
#> 133   2/9/22       1560  1350.837   209.163281
#> 134  2/10/22       1710  1352.665   357.334893
#> 135  2/11/22       1770  1356.113   413.887031
#> 136  2/12/22       1800  1361.349   438.651028
#> 137  2/13/22       1860  1368.339   491.661373
#> 138  2/14/22       1800  1375.958   424.041762
#> 139  2/15/22       1860  1382.948   477.052264
#> 140  2/16/22       1900  1389.093   510.906743
#> 141  2/17/22       1930  1394.270   535.729516
#> 142  2/18/22       1840  1398.351   441.648921
#> 143  2/19/22       1810  1401.222   408.778290
#> 144  2/20/22       1760  1402.794   357.206285
#> 145  2/21/22       1790  1396.581   393.418532
#> 146  2/22/22       1800  1376.411   423.588510
#> 147  2/23/22       1820  1353.650   466.349567
#> 148  2/24/22       1770  1328.160   441.840115
#> 149  2/25/22       1630  1301.944   328.056094
#> 150  2/26/22       1470  1281.831   188.168750
#> 151  2/27/22       1330  1272.250    57.750000
#> 152  2/28/22       1270  1270.000     0.000000
#> 153   3/1/22       1500  1270.323   229.676563
#> 154   3/2/22       1820  1272.019   547.981328
#> 155   3/3/22       1930  1276.256   653.743761
#> 156   3/4/22       1890  1283.348   606.651621
#> 157   3/5/22       1930  1292.942   637.058494
#> 158   3/6/22       2010  1304.789   705.210670
#> 159   3/7/22       2100  1318.907   781.092503
#> 160   3/8/22       2360  1335.620  1024.379939
#> 161   3/9/22       2850  1356.005  1493.994677
#> 162  3/10/22       2820  1381.120  1438.880066
#> 163  3/11/22       2490  1410.280  1079.719553
#> 164  3/12/22       2330  1441.543   888.456603
#> 165  3/13/22       2270  1473.431   796.569049
#> 166  3/14/22       2090  1505.027   584.973309
#> 167  3/15/22       1960  1535.422   424.578417
#> 168  3/16/22       1960  1563.999   396.000721
#> 169  3/17/22       1960  1590.679   369.321426
#> 170  3/18/22       1990  1615.626   374.374334
#> 171  3/19/22       2080  1639.161   440.838747
#> 172  3/20/22       2360  1662.046   697.954206
#> 173  3/21/22       2660  1685.580   974.419502
#> 174  3/22/22       2690  1710.819   979.181199
#> 175  3/23/22       2750  1737.963  1012.036595
#> 176  3/24/22       2710  1766.741   943.258912
#> 177  3/25/22       2640  1796.628   843.372363
#> 178  3/26/22       2560  1826.929   733.070777
#> 179  3/27/22       2440  1856.923   583.076710
#> 180  3/28/22       2370  1885.937   484.062917
#> 181  3/29/22       2530  1913.808   616.191700
#> 182  3/30/22       2610  1941.009   668.990810
#> 183  3/31/22       2620  1967.948   652.052414
#> 184   4/1/22       2820  1994.932   825.068266
#> 185   4/2/22       3140  2022.830  1117.170196
#> 186   4/3/22       3450  2052.970  1397.029889
#> 187   4/4/22       3520  2086.280  1433.720296
#> 188   4/5/22       3380  2122.550  1257.450412
#> 189   4/6/22       3120  2160.371   959.628773
#> 190   4/7/22       2950  2197.003   752.996566
#> 191   4/8/22       3010  2230.700   779.299777
#> 192   4/9/22       3160  2261.511   898.488690
#> 193  4/10/22       3120  2289.457   830.543441
#> 194  4/11/22       3110  2314.529   795.470695
#> 195  4/12/22       2990  2336.726   653.274499
#> 196  4/13/22       2780  2356.124   423.875653
#> 197  4/14/22       2660  2373.084   286.915826
#> 198  4/15/22       2740  2388.112   351.887751
#> 199  4/16/22       2940  2401.453   538.546912
#> 200  4/17/22       2980  2413.111   566.889455
#> 201  4/18/22       2840  2423.046   416.953598
#> 202  4/19/22       2720  2431.233   288.767298
#> 203  4/20/22       2610  2437.845   172.155253
#> 204  4/21/22       2510  2443.397    66.602984
#> 205  4/22/22       2550  2448.448   101.551510
#> 206  4/23/22       2650  2453.426   196.573741
#> 207  4/24/22       2700  2458.721   241.279222
#> 208  4/25/22       2830  2464.721   365.279404
#> 209  4/26/22       3540  2472.725  1067.275488
#> 210  4/27/22       4250  2485.576  1764.423712
#> 211  4/28/22       4920  2506.441  2413.558778
#> 212  4/29/22       4860  2536.843  2323.156514
#> 213  4/30/22       4570  2575.601  1994.399258
#> 214   5/1/22       4290  2619.995  1670.005023
#> 215   5/2/22       4090  2667.486  1422.513928
#> 216   5/3/22       4330  2716.741  1613.258968
#> 217   5/4/22       4850  2768.353  2081.646862
#> 218   5/5/22       5030  2823.745  2206.255115
#> 219   5/6/22       4900  2883.114  2016.886303
#> 220   5/7/22       5180  2945.833  2234.167316
#> 221   5/8/22       5390  3011.965  2378.034629
#> 222   5/9/22       5570  3081.884  2488.115592
#> 223  5/10/22       5810  3155.789  2654.211090
#> 224  5/11/22       6240  3234.221  3005.779426
#> 225  5/12/22       8100  3320.247  4779.752887
#> 226  5/13/22       8950  3419.319  5530.680631
#> 227  5/14/22       9000  3534.073  5465.926854
#> 228  5/15/22       8920  3662.822  5257.177628
#> 229  5/16/22      10100  3804.327  6295.673072
#> 230  5/17/22      10300  3959.437  6340.563200
#> 231  5/18/22       9660  4126.637  5533.362898
#> 232  5/19/22      10200  4302.482  5897.517598
#> 233  5/20/22      11200  4486.760  6713.239552
#> 234  5/21/22      12000  4681.914  7318.085926
#> 235  5/22/22      12800  4890.057  7909.943372
#> 236  5/23/22      13500  5106.436  8393.563934
#> 237  5/24/22      13200  5314.897  7885.103472
#> 238  5/25/22      11300  5504.409  5795.590876
#> 239  5/26/22       9540  5674.468  3865.532349
#> 240  5/27/22       8270  5826.202  2443.797874
#> 241  5/28/22       7640  5963.123  1676.877096
#> 242  5/29/22      10400  6092.770  4307.230064
#> 243  5/30/22      12800  6227.421  6572.578903
#> 244  5/31/22      14300  6365.745  7934.255057
#> 245   6/1/22      16100  6494.829  9605.171405
#> 246   6/2/22      17100  6608.862 10491.138226
#> 247   6/3/22      16900  6705.172 10194.828046
#> 248   6/4/22      15700  6781.060  8918.939750
#> 249   6/5/22      13800  6834.241  6965.759315
#> 250   6/6/22      10900  6863.322  4036.677757
#> 251   6/7/22       9860  6811.700  3048.299564
#> 252   6/8/22       7910  6643.595  1266.404934
#> 253   6/9/22       7820  6544.562  1275.437766
#> 254  6/10/22       8730  6439.765  2290.235085
#> 255  6/11/22       9400  6321.405  3078.594644
#> 256  6/12/22       9550  6185.115  3364.884847
#> 257  6/13/22       9290  6028.986  3261.013876
#> 258  6/14/22       9300  5852.616  3447.383868
#> 259  6/15/22       9700  5654.689  4045.310586
#> 260  6/16/22       9860  5432.528  4427.472301
#> 261  6/17/22       9600  5184.081  4415.919320
#> 262  6/18/22       8730  4909.707  3820.292538
#> 263  6/19/22       7470  4656.659  2813.340965
#> 264  6/20/22       6110  4483.685  1626.314557
#> 265  6/21/22       5680  4369.255  1310.745467
#> 266  6/22/22       5530  4269.059  1260.941045
#> 267  6/23/22       5700  4166.562  1533.437954
#> 268  6/24/22       5800  4054.847  1745.153405
#> 269  6/25/22       6630  3931.409  2698.591072
#> 270  6/26/22       7970  3790.786  4179.214428
#> 271  6/27/22       7600  3627.348  3972.652264
#> 272  6/28/22       6540  3440.801  3099.199134
#> 273  6/29/22       5130  3269.680  1860.319971
#> 274  6/30/22       4520  3143.573  1376.426995
#> 275   7/1/22       4080  3049.809  1030.191346
#> 276   7/2/22       4280  2965.976  1314.024428
#> 277   7/3/22       4030  2877.372  1152.627761
#> 278   7/4/22       3730  2796.078   933.921903
#> 279   7/5/22       3710  2721.166   988.834490
#> 280   7/6/22       4430  2639.895  1790.104854
#> 281   7/7/22       4360  2548.940  1811.059977
#> 282   7/8/22       3660  2453.732  1206.268219
#> 283   7/9/22       3310  2370.115   939.884561
#> 284  7/10/22       3190  2298.773   891.226552
#> 285  7/11/22       3070  2231.377   838.623300
#> 286  7/12/22       2940  2168.650   771.349513
#> 287  7/13/22       2780  2112.595   667.404879
#> 288  7/14/22       2620  2064.968   555.032302
#> 289  7/15/22       2520  2024.019   495.980867
#> 290  7/16/22       2460  1986.237   473.763099
#> 291  7/17/22       2370  1951.472   418.527675
#> 292  7/18/22       2210  1924.024   285.975865
#> 293  7/19/22       2070  1906.513   163.487421
#> 294  7/20/22       1990  1896.500    93.499915
#> 295  7/21/22       1960  1890.135    69.864773
#> 296  7/22/22       1930  1885.687    44.313268
#> 297  7/23/22       1940  1882.079    57.921438
#> 298  7/24/22       1990  1878.057   111.942645
#> 299  7/25/22       1990  1873.431   116.568570
#> 300  7/26/22       2040  1868.021   171.979371
#> 301  7/27/22       2030  1861.655   168.344777
#> 302  7/28/22       2060  1854.205   205.794711
#> 303  7/29/22       2070  1845.534   224.465695
#> 304  7/30/22       2100  1835.468   264.532078
#> 305  7/31/22       1970  1824.390   145.609507
#> 306   8/1/22       1930  1814.206   115.794062
#> 307   8/2/22       2410  1804.088   605.912500
#> 308   8/3/22       2090  1791.501   298.499324
#> 309   8/4/22       1940  1778.218   161.782377
#> 310   8/5/22       1920  1765.911   154.089056
#> 311   8/6/22       1800  1758.282    41.717899
#> 312   8/7/22       1810  1754.884    55.115634
#> 313   8/8/22       2080  1750.756   329.243802
#> 314   8/9/22       1890  1745.325   144.674804
#> 315  8/10/22       1770  1741.125    28.875000
#> 316  8/11/22       1740  1740.000     0.000000
#> 317  8/12/22       1760  1740.007    19.993007
#> 318  8/13/22       1760  1740.017    19.983453
#> 319  8/14/22       1740  1740.000     0.000000
#> 320  8/15/22       2500  1741.069   758.931250
#> 321  8/16/22       1930  1744.382   185.618125
#> 322  8/17/22       1960  1749.084   210.915871
#> 323  8/18/22       1920  1753.799   166.200621
#> 324  8/19/22       1870  1758.105   111.895465
#> 325  8/20/22       1810  1762.022    47.978393
#> 326  8/21/22       2050  1765.957   284.042513
#> 327  8/22/22       2030  1770.557   259.442762
#> 328  8/23/22       1910  1775.584   134.416128
#> 329  8/24/22       2020  1780.352   239.648265
#> 330  8/25/22       2050  1784.661   265.338918
#> 331  8/26/22       2030  1788.464   241.536223
#> 332  8/27/22       1980  1791.709   188.290950
#> 333  8/28/22       1900  1794.363   105.636896
#> 334  8/29/22       1870  1796.453    73.546822
#> 335  8/30/22       1890  1798.058    91.941695
#> 336  8/31/22       1900  1799.220   100.780237
#> 337   9/1/22       1870  1799.942    70.058150
#> 338   9/2/22       1830  1800.295    29.705186
#> 339   9/3/22       1830  1800.406    29.593673
#> 340   9/4/22       1840  1798.290    41.710080
#> 341   9/5/22       1800  1796.920     3.080154
#> 342   9/6/22       1850  1796.594    53.405910
#> 343   9/7/22       1840  1796.035    43.965215
#> 344   9/8/22       1830  1795.208    34.791876
#> 345   9/9/22       1850  1794.124    55.875718
#> 346  9/10/22       1860  1792.716    67.284142
#> 347  9/11/22       1830  1790.959    39.040915
#> 348  9/12/22       1790  1790.000     0.000000
#> 349  9/13/22       1840  1790.070    49.929688
#> 350  9/14/22       2090  1790.693   299.307422
#> 351  9/15/22       2530  1793.176   736.824204
#> 352  9/16/22       2660  1799.009   860.990802
#> 353  9/17/22       2920  1809.026  1110.973712
#> 354  9/18/22       2680  1823.143   856.856737
#> 355  9/19/22       2580  1840.239   739.761206
#> 356  9/20/22       2450  1858.985   591.014873
#> 357  9/21/22       2440  1878.518   561.481834
#> 358  9/22/22       2850  1898.980   951.019916
#> 359  9/23/22       2700  1921.050   778.950201
#> 360  9/24/22       2550  1944.315   605.685206
#> 361  9/25/22       2610  1967.923   642.076803
#> 362  9/26/22       2540  1991.552   548.447994
#> 363  9/27/22       2490  2014.883   475.117449
#> 364  9/28/22       2420  2037.489   382.510841
#> 365  9/29/22       2370  2059.012   310.988313
#> 366  9/30/22       2390  2079.275   310.725354
#> 367  10/1/22       2830  2098.951   731.049404
#> 368  10/2/22       3410  2120.095  1289.905141
#> 369  10/3/22       2840  2143.825   696.175053
#> 370  10/4/22       2810  2168.804   641.195636
#> 371  10/5/22       2880  2193.772   686.228450
#> 372  10/6/22       3040  2217.647   822.352727
#> 373  10/7/22       3650  2239.460  1410.539760
#> 374  10/8/22       3290  2259.013  1030.986717
#> 375  10/9/22       2930  2276.089   653.911125
#> 376 10/10/22       2760  2290.578   469.422288
#> 377 10/11/22       2650  2302.589   347.410771
#> 378 10/12/22       2570  2312.507   257.492721
#> 379 10/13/22       2480  2320.875   159.125304
#> 380 10/14/22       2460  2328.168   131.831959
#> 381 10/15/22       2450  2334.645   115.355497
#> 382 10/16/22       2430  2340.435    89.565320
#> 383 10/17/22       2410  2345.680    64.319796
#> 384 10/18/22       2430  2350.532    79.467686
#> 385 10/19/22       2450  2355.131    94.869094
#> 386 10/20/22       2430  2359.543    70.457035
#> 387 10/21/22       2430  2363.742    66.257521
#> 388 10/22/22       2440  2367.722    72.277551
#> 389 10/23/22       2730  2371.928   358.071878
#> 390 10/24/22       2480  2376.781   103.218620
#> 391 10/25/22       2810  2382.330   427.669821
#> 392 10/26/22       2930  2389.188   540.811799
#> 393 10/27/22       2790  2397.733   392.267463
#> 394 10/28/22       2720  2407.201   312.798975
#> 395 10/29/22       2640  2416.453   223.546981
#> 396 10/30/22       2620  2424.926   195.074317
#> 397 10/31/22       2660  2432.625   227.375246
#> 398  11/1/22       2660  2439.556   220.444162
#> 399  11/2/22       2630  2445.719   184.281028
#> 400  11/3/22       2640  2451.115   188.885039
#> 401  11/4/22       2670  2455.744   214.255593
#> 402  11/5/22       2680  2459.600   220.400036
#> 403  11/6/22       2720  2462.663   257.337448
#> 404  11/7/22       2720  2464.901   255.098805
#> 405  11/8/22       2700  2466.277   233.722644
#> 406  11/9/22       2720  2466.751   253.248835
#> 407 11/10/22       2760  2453.986   306.014283
#> 408 11/11/22       2750  2439.272   310.728374
#> 409 11/12/22       2730  2422.683   307.316828
#> 410 11/13/22       2800  2404.089   395.910923
#> 411 11/14/22       2770  2383.239   386.760559
#> 412 11/15/22       2740  2360.038   379.962116
#> 413 11/16/22       2710  2334.526   375.474255
#> 414 11/17/22       2700  2306.700   393.300087
#> 415 11/18/22       2710  2276.452   433.548016
#> 416 11/19/22       2660  2244.513   415.487361
#> 417 11/20/22       2550  2215.284   334.716066
#> 418 11/21/22       2470  2191.388   278.611963
#> 419 11/22/22       2400  2171.636   228.364285
#> 420 11/23/22       2290  2157.579   132.420848
#> 421 11/24/22       2340  2146.766   193.233687
#> 422 11/25/22       2450  2134.687   315.312812
#> 423 11/26/22       2620  2120.599   499.400596
#> 424 11/27/22       2570  2103.809   466.190816
#> 425 11/28/22       2570  2084.108   485.891994
#> 426 11/29/22       2660  2061.315   598.684744
#> 427 11/30/22       2570  2035.156   534.844497
#> 428  12/1/22       2450  2005.789   444.211074
#> 429  12/2/22       2540  1973.402   566.598191
#> 430  12/3/22       2520  1937.736   582.263609
#> 431  12/4/22       2300  1903.587   396.412942
#> 432  12/5/22       2350  1871.369   478.630545
#> 433  12/6/22       2650  1835.861   814.138807
#> 434  12/7/22       2720  1795.752   924.248252
#> 435  12/8/22       2660  1750.220   909.780489
#> 436  12/9/22       2680  1699.033   980.966797
#> 437 12/10/22       2620  1642.003   977.996822
#> 438 12/11/22       2570  1579.011   990.989328
#> 439 12/12/22       2530  1509.977  1020.023323
#> 440 12/13/22       1910  1456.255   453.745018
#> 441 12/14/22       1710  1427.573   282.427047
#> 442 12/15/22       1650  1407.106   242.894105
#> 443 12/16/22       1590  1389.844   200.155789
#> 444 12/17/22       1520  1376.453   143.546799
#> 445 12/18/22       1450  1367.652    82.347891
#> 446 12/19/22       1410  1362.597    47.403125
#> 447 12/20/22       1370  1360.375     9.625000
#> 448 12/21/22       1360  1360.000     0.000000
#> 449 12/22/22       1380  1360.028    19.971875
#> 450 12/23/22       1450  1360.235    89.765156
#> 451 12/24/22       1710  1361.184   348.816166
#> 452 12/25/22       2000  1364.000   635.999908
#> 453 12/26/22       2260  1369.945   890.054935
#> 454 12/27/22       2460  1379.954  1080.046083
#> 455 12/28/22       2650  1394.579  1255.421063
#> 456 12/29/22       2810  1414.112  1395.887974
#> 457 12/30/22       2790  1438.424  1351.576230
#> 458 12/31/22       2850  1466.941  1383.059378
#> 459   1/1/23       2970  1499.204  1470.795562
#> 460   1/2/23       3020  1534.985  1485.015047
#> 461   1/3/23       2930  1573.756  1356.244010
#> 462   1/4/23       2800  1614.502  1185.497918
#> 463   1/5/23       2840  1656.273  1183.727055
#> 464   1/6/23       2930  1698.742  1231.258083
#> 465   1/7/23       2800  1741.696  1058.304117
#> 466   1/8/23       2690  1784.429   905.570545
#> 467   1/9/23       2610  1826.129   783.870610
#> 468  1/10/23       2750  1866.528   883.472456
#> 469  1/11/23       2780  1905.908   874.091940
#> 470  1/12/23       2720  1944.358   775.642306
#> 471  1/13/23       2730  1980.819   749.181134
#> 472  1/14/23       2750  2014.404   735.596117
#> 473  1/15/23       2890  2045.179   844.820970
#> 474  1/16/23       2820  2073.180   746.820094
#> 475  1/17/23       2790  2098.425   691.574794
#> 476  1/18/23       2720  2120.941   599.058653
#> 477  1/19/23       2750  2140.767   609.232934
#> 478  1/20/23       2720  2157.942   562.057633
#> 479  1/21/23       2700  2172.502   527.497694
#> 480  1/22/23       2620  2184.483   435.517356
#> 481  1/23/23       2570  2193.978   376.022404
#> 482  1/24/23       2470  2201.289   268.711420
#> 483  1/25/23       2340  2207.011   132.988884
#> 484  1/26/23       2270  2211.834    58.166280
#> 485  1/27/23       2330  2216.281   113.719434
#> 486  1/28/23       2500  2220.880   279.120242
#> 487  1/29/23       2570  2226.244   343.755632
#> 488  1/30/23       2540  2232.112   307.888334
#> 489  1/31/23       2510  2237.670   272.330023
#> 490   2/1/23       2440  2242.637   197.363169
#> 491   2/2/23       2390  2246.980   143.020313
#> 492   2/3/23       2350  2250.696    99.303893
#> 493   2/4/23       2320  2253.852    66.148328
#> 494   2/5/23       2290  2256.605    33.395016
#> 495   2/6/23       2290  2259.110    30.890390
#> 496   2/7/23       2310  2261.455    48.545485
#> 497   2/8/23       2330  2263.734    66.266058
#> 498   2/9/23       2360  2266.071    93.928915
#> 499  2/10/23       2400  2268.613   131.386783
#> 500  2/11/23       2430  2271.513   158.486684
#> 501  2/12/23       2480  2274.914   205.085674
#> 502  2/13/23       2520  2278.964   241.036140
#> 503  2/14/23       2530  2283.742   246.257617
#> 504  2/15/23       2510  2289.174   220.825982
#> 505  2/16/23       2480  2295.050   184.950455
#> 506  2/17/23       2480  2301.159   178.840674
#> 507  2/18/23       2480  2307.393   172.607014
#> 508  2/19/23       2490  2313.712   176.288361
#> 509  2/20/23       2520  2320.138   199.862217
#> 510  2/21/23       2570  2326.789   243.211497
#> 511  2/22/23       2700  2333.960   366.040411
#> 512  2/23/23       2730  2342.014   387.986297
#> 513  2/24/23       2840  2351.200   488.800074
#> 514  2/25/23       2840  2361.655   478.345298
#> 515  2/26/23       2820  2373.263   446.737364
#> 516  2/27/23       2650  2385.497   264.503364
#> 517  2/28/23       2740  2397.818   342.182192
#> 518   3/1/23       2790  2409.884   380.115801
#> 519   3/2/23       2820  2421.236   398.764018
#> 520   3/3/23       2870  2431.567   438.433114
#> 521   3/4/23       2860  2440.825   419.174690
#> 522   3/5/23       2850  2448.954   401.046267
#> 523   3/6/23       2670  2455.907   214.092873
#> 524   3/7/23       2570  2461.802   108.198463
#> 525   3/8/23       2530  2466.973    63.027328
#> 526   3/9/23       2550  2471.728    78.272153
#> 527  3/10/23       2580  2476.251   103.749164
#> 528  3/11/23       2690  2480.817   209.183092
#> 529  3/12/23       2680  2485.732   194.268341
#> 530  3/13/23       2640  2490.987   149.012586
#> 531  3/14/23       2700  2496.463   203.536810
#> 532  3/15/23       2980  2502.603   477.397329
#> 533  3/16/23       3260  2510.541   749.458876
#> 534  3/17/23       3190  2521.057   668.943019
#> 535  3/18/23       3660  2534.577  1125.423209
#> 536  3/19/23       4380  2552.827  1827.173253
#> 537  3/20/23       4330  2577.637  1752.363098
#> 538  3/21/23       4180  2608.582  1571.418491
#> 539  3/22/23       4190  2644.123  1545.877283
#> 540  3/23/23       4020  2682.975  1337.024902
#> 541  3/24/23       4210  2724.245  1485.754819
#> 542  3/25/23       4640  2766.838  1873.161562
#> 543  3/26/23       5100  2807.792  2292.208335
#> 544  3/27/23       4450  2844.753  1605.246535
#> 545  3/28/23       3730  2877.418   852.581971
#> 546  3/29/23       3320  2906.149   413.850633
#> 547  3/30/23       3250  2932.037   317.963398
#> 548  3/31/23       3310  2955.968   354.031768
#> 549   4/1/23       3340  2978.394   361.606026
#> 550   4/2/23       3300  2999.518   300.482404
#> 551   4/3/23       3330  3019.380   310.619854
#> 552   4/4/23       3430  3038.221   391.779286
#> 553   4/5/23       3780  3056.896   723.103503
#> 554   4/6/23       4290  3077.168  1212.831580
#> 555   4/7/23       4070  3100.309   969.690800
#> 556   4/8/23       3890  3125.620   764.380210
#> 557   4/9/23       3880  3151.815   728.185135
#> 558  4/10/23       3800  3178.226   621.774370
#> 559  4/11/23       3830  3204.476   625.524491
#> 560  4/12/23       4140  3230.848   909.152112
#> 561  4/13/23       4930  3259.202  1670.798266
#> 562  4/14/23       7490  3295.349  4194.650703
#> 563  4/15/23      10900  3351.068  7548.931722
#> 564  4/16/23      14700  3441.754 11258.245990
#> 565  4/17/23      17400  3581.128 13818.872252
#> 566  4/18/23      15700  3771.923 11928.076942
#> 567  4/19/23      12400  4000.018  8399.982271
#> 568  4/20/23      10500  4232.866  6267.133624
#> 569  4/21/23      10600  4447.674  6152.325919
#> 570  4/22/23      12600  4642.876  7957.123536
#> 571  4/23/23      12700  4817.869  7882.131077
#> 572  4/24/23      10100  4971.842  5128.158305
#> 573  4/25/23       8420  5104.667  3315.333116
#> 574  4/26/23       7190  5219.024  1970.976140
#> 575  4/27/23       6560  5320.483  1239.516992
#> 576  4/28/23       6560  5413.447  1146.553218
#> 577  4/29/23       8020  5501.491  2518.508602
#> 578  4/30/23       9870  5591.540  4278.460378
#> 579   5/1/23      10700  5691.219  5008.780903
#> 580   5/2/23      11800  5805.061  5994.938571
#> 581   5/3/23      12000  5934.924  6065.076322
#> 582   5/4/23      13500  6081.981  7418.018756
#> 583   5/5/23      16100  6251.081  9848.919395
#> 584   5/6/23      18400  6450.745 11949.255458
#> 585   5/7/23      20200  6688.093 13511.906565
#> 586   5/8/23      21100  6965.914 14134.086069
#> 587   5/9/23      20900  7281.581 13618.418547
#> 588  5/10/23      19500  7626.591 11873.409420
#> 589  5/11/23      17500  7987.733  9512.266808
#> 590  5/12/23      16800  8352.070  8447.929534
#> 591  5/13/23      18200  8714.279  9485.720726
#> 592  5/14/23      20300  9078.536 11221.464261
#> 593  5/15/23      21500  9452.058 12047.942211
#> 594  5/16/23      22100  9838.579 12261.421482
#> 595  5/17/23      22100 10237.422 11862.577937
#> 596  5/18/23      23200 10646.956 12553.043554
#> 597  5/19/23      24800 11068.678 13731.321702
#> 598  5/20/23      26100 11506.331 14593.669133
#> 599  5/21/23      26800 11962.043 14837.957014
#> 600  5/22/23      27200 12435.003 14764.996700
#> 601  5/23/23      27600 12922.733 14677.266924
#> 602  5/24/23      27900 13420.323 14479.677011
#> 603  5/25/23      27800 13901.227 13898.772660
#> 604  5/26/23      27900 14342.073 13557.927281
#> 605  5/27/23      28000 14742.171 13257.828810
#> 606  5/28/23      27900 15100.817 12799.183114
#> 607  5/29/23      28000 15417.276 12582.724120
#> 608  5/30/23      27200 15690.816 11509.183776
#> 609  5/31/23      26100 15920.832 10179.167661
#> 610   6/1/23      25400 16106.983  9293.016687
#> 611   6/2/23      24400 16249.202  8150.797569
#> 612   6/3/23      23800 16347.633  7452.366600
#> 613   6/4/23      22800 16403.640  6396.359834
#> 614   6/5/23      21200 16422.096  4777.903753
#> 615   6/6/23      21000 16063.534  4936.466128
#> 616   6/7/23      20100 15699.766  4400.233652
#> 617   6/8/23      19500 15367.315  4132.685029
#> 618   6/9/23      20500 15030.712  5469.287869
#> 619  6/10/23      23300 14658.114  8641.886379
#> 620  6/11/23      24100 14236.004  9863.996360
#> 621  6/12/23      24400 13754.670 10645.329967
#> 622  6/13/23      24300 13209.209 11090.790581
#> 623  6/14/23      23600 12597.217 11002.783030
#> 624  6/15/23      21800 11919.988  9880.011997
#> 625  6/16/23      19900 11227.640  8672.360298
#> 626  6/17/23      18000 10601.502  7398.497620
#> 627  6/18/23      16200 10074.597  6125.402832
#> 628  6/19/23      15800  9594.159  6205.840900
#> 629  6/20/23      16000  9090.679  6909.321243
#> 630  6/21/23      14900  8582.862  6317.137830
#> 631  6/22/23      14400  8090.932  6309.067925
#> 632  6/23/23      14000  7595.602  6404.397756
#> 633  6/24/23      13300  7104.705  6195.294872
#> 634  6/25/23      12200  6646.979  5553.021483
#> 635  6/26/23      11400  6229.166  5170.834036
#> 636  6/27/23      10600  5842.342  4757.658417
#> 637  6/28/23       9710  5492.667  4217.333424
#> 638  6/29/23       8730  5190.450  3539.549647
#> 639  6/30/23       7980  4933.865  3046.134754
#> 640   7/1/23       7460  4707.962  2752.037572
#> 641   7/2/23       6910  4507.122  2402.878456
#> 642   7/3/23       6350  4334.996  2015.003736
#> 643   7/4/23       5950  4187.834  1762.166201
#> 644   7/5/23       5710  4054.685  1655.314812
#> 645   7/6/23       5520  3928.173  1591.826824
#> 646   7/7/23       5300  3808.025  1491.974945
#> 647   7/8/23       5120  3694.351  1425.648589
#> 648   7/9/23       4920  3586.866  1333.133610
#> 649  7/10/23       4780  3484.450  1295.549848
#> 650  7/11/23       4610  3386.297  1223.702539
#> 651  7/12/23       4380  3296.403  1083.597339
#> 652  7/13/23       4190  3216.246   973.753880
#> 653  7/14/23       4000  3144.996   855.004195
#> 654  7/15/23       3780  3084.590   695.409940
#> 655  7/16/23       3610  3035.097   574.902638
#> 656  7/17/23       3480  2993.754   486.246095
#> 657  7/18/23       3310  2961.220   348.779563
#> 658  7/19/23       3190  2937.806   252.194122
#> 659  7/20/23       3120  2920.196   199.804456
#> 660  7/21/23       3030  2907.644   122.356168
#> 661  7/22/23       2980  2899.750    80.249912
#> 662  7/23/23       2930  2895.270    34.729634
#> 663  7/24/23       2960  2892.409    67.591159
#> 664  7/25/23       3000  2889.121   110.878913
#> 665  7/26/23       2980  2885.250    94.749715
#> 666  7/27/23       3010  2880.727   129.272711
#> 667  7/28/23       3100  2875.327   224.673041
#> 668  7/29/23       2990  2868.864   121.135572
#> 669  7/30/23       2950  2861.940    88.060218
#> 670  7/31/23       2930  2855.611    74.389425
#> 671   8/1/23       2930  2849.579    80.421000
#> 672   8/2/23       3020  2842.922   177.078446
#> 673   8/3/23       3090  2835.218   254.782314
#> 674   8/4/23       3220  2825.873   394.126751
#> 675   8/5/23       3160  2814.421   345.579054
#> 676   8/6/23       3240  2800.654   439.345688
#> 677   8/7/23       3060  2784.611   275.388749
#> 678   8/8/23       3010  2766.696   243.304107
#> 679   8/9/23       2970  2748.723   221.276631
#> 680  8/10/23       2980  2730.767   249.233182
#> 681  8/11/23       2990  2711.295   278.705341
#> 682  8/12/23       2950  2691.070   258.930309
#> 683  8/13/23       2920  2671.292   248.708442
#> 684  8/14/23       2870  2653.153   216.846965
#> 685  8/15/23       2820  2637.598   182.402124
#> 686  8/16/23       2800  2623.619   176.380675
#> 687  8/17/23       2720  2612.561   107.438567
#> 688  8/18/23       2650  2606.688    43.311964
#> 689  8/19/23       2850  2602.872   247.127799
#> 690  8/20/23       2960  2597.690   362.309614
#> 691  8/21/23       2800  2590.716   209.284190
#> 692  8/22/23       2800  2582.225   217.774683
#> 693  8/23/23       2640  2575.128    64.871522
#> 694  8/24/23       2740  2569.717   170.283402
#> 695  8/25/23       2680  2563.512   116.487546
#> 696  8/26/23       2590  2558.984    31.015675
#> 697  8/27/23       2590  2556.470    33.530459
#> 698  8/28/23       2610  2553.720    56.279550
#> 699  8/29/23       2650  2550.599    99.401288
#> 700  8/30/23       2700  2546.857   153.142695
#> 701  8/31/23       2610  2542.397    67.602537
#> 702   9/1/23       2540  2540.000     0.000000
#> 703   9/2/23       2550  2540.014     9.985938
#> 704   9/3/23       2640  2540.195    99.805234
#> 705   9/4/23       2620  2540.683    79.316698
#> 706   9/5/23       2670  2541.286   128.713682
#> 707   9/6/23       2700  2541.784   158.215725
#> 708   9/7/23       2750  2542.123   207.877465
#> 709   9/8/23       2760  2541.944   218.055944
#> 710   9/9/23       2760  2537.338   222.661783
#> 711  9/10/23       2730  2531.424   198.576039
#> 712  9/11/23       2650  2524.378   125.621839
#> 713  9/12/23       2600  2517.391    82.609344
#> 714  9/13/23       2620  2510.662   109.337805
#> 715  9/14/23       2880  2502.904   377.096022
#> 716  9/15/23       2640  2493.613   146.387256
#> 717  9/16/23       2580  2484.863    95.136816
#> 718  9/17/23       2630  2477.073   152.926625
#> 719  9/18/23       2610  2468.460   141.539975
#> 720  9/19/23       2590  2458.986   131.014041
#> 721  9/20/23       2560  2449.746   110.253529
#> 722  9/21/23       2530  2442.023    87.976788
#> 723  9/22/23       2540  2434.875   105.125244
#> 724  9/23/23       2570  2427.057   142.943380
#> 725  9/24/23       2610  2418.354   191.645969
#> 726  9/25/23       2720  2408.380   311.620392
#> 727  9/26/23       2830  2396.510   433.489804
#> 728  9/27/23       2810  2382.202   427.797667
#> 729  9/28/23       2760  2365.339   394.661395
#> 730  9/29/23       2740  2346.029   393.970698
#> 731  9/30/23       2740  2324.294   415.706175
#> 732  10/1/23       2760  2300.000   460.000000